fixed scrollview rebase style issues
