+++ src/theme/components/booking/groupBookingOptions.js
@@ -7,9 +7,8 @@ export const groupBookingOptions = {
         marginTop: 12,
         borderColor: colors.lightGray02,
         borderWidth: 1,
-        overflow: 'auto',
         flex: 1,
-        padding: 12,
+        padding: 10,
         paddingTop: 14,
         paddingBottom: 14,
       },
